+++ AngularClient/src/app-settings/app-settings.ts
@@ -37,5 +37,5 @@ export class AppSettings {
   /**
    * Application version number
    */
-  static get version(): string { return "0.86.3"; }
+  static get version(): string { return "0.86.4"; }
 }