+++ AngularClient/src/main/core/project.service.ts
@@ -125,7 +125,7 @@ export class ProjectService {
     return elementCell;
   }
 
-  createElementField(initialValues: Object) {
+  createElementField(initialValues: Object, rating: number = 50) {
 
     const elementField = this.appEntityManager.createEntity("ElementField", initialValues) as ElementField;
 
@@ -138,7 +138,7 @@ export class ProjectService {
       const userElementFieldInitial = {
         User: this.authService.currentUser,
         ElementField: elementField,
-        Rating: 0
+        Rating: rating
       };
 
       this.appEntityManager.createEntity("UserElementField", userElementFieldInitial) as UserElementField;